+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## 1.0.0 2017-03-15
+### Added
+- Publications e.g. Doctoral/Master's Thesis.
+
+### Changed
+- Configuration style to match Puree v1.0.0.
+
 ## 0.1.0 - 2017-02-03
 ### Added
 - Working product supports datasets.